260110_Backup_v1_배포직전/requirements.txt
streamlit==1.40.1
streamlit-authenticator==0.3.3
pandas==2.2.3
pyyaml==6.0.2
gspread==6.1.4
google-auth==2.35.0
